+++ plugins/jira-dashboard/src/components/JiraDashboardContent/JiraDashboardContent.test.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { JiraDashboardContent } from './JiraDashboardContent';
 import { rest } from 'msw';
 import { setupServer } from 'msw/node';
@@ -14,6 +13,7 @@ import { ApiProvider, UrlPatternDiscovery } from '@backstage/core-app-api';
 import mockedJiraResponse from '../../../dev/__fixtures__/jiraResponse.json';
 import mockedEntity from '../../../dev/__fixtures__/entity.json';
 import singleProjectResponse from '../../../dev/__fixtures__/singleProjectResponse.json';
+import { fireEvent, waitFor } from '@testing-library/react';
 
 describe('JiraDashboardContent', () => {
   const server = setupServer();
@@ -134,5 +134,61 @@ describe('JiraDashboardContent - Single Project View', () => {
     );
 
     expect(getAllByTestId('issue-table')).toHaveLength(2); // Matches 2 filters in the JSON
+  });
+});
+describe('JiraDashboardContent - Multi-project response', () => {
+  const server = setupServer();
+  registerMswTestHooks(server);
+
+  const mockBaseUrl = 'http://localhost:7007/api/jira-dashboard';
+  const discoveryApi = UrlPatternDiscovery.compile(mockBaseUrl);
+  const fetchApi = new MockFetchApi();
+  let jiraClient: JiraDashboardClient;
+  let apis: TestApiRegistry;
+
+  const renderComponent = async () =>
+    await renderInTestApp(
+      <EntityProvider entity={mockedEntity}>
+        <ApiProvider apis={apis}>
+          <JiraDashboardContent />
+        </ApiProvider>
+      </EntityProvider>,
+    );
+
+  beforeEach(() => {
+    server.use(
+      rest.get(
+        `${mockBaseUrl}/dashboards/by-entity-ref/:kind/:namespace/:name`,
+        (_req, res, ctx) => res(ctx.json(mockedJiraResponse)),
+      ),
+    );
+    jiraClient = new JiraDashboardClient({ discoveryApi, fetchApi });
+    apis = TestApiRegistry.from([jiraDashboardApiRef, jiraClient]);
+  });
+
+  it('renders the dashboard title', async () => {
+    const { getByText } = await renderComponent();
+    expect(getByText(/Jira Dashboard/i)).toBeInTheDocument();
+  });
+
+  it('renders 1 project card and 2 tables per tab', async () => {
+    const { getAllByTestId, getAllByRole } = await renderComponent();
+
+    const tabs = getAllByRole('tab');
+    expect(tabs).toHaveLength(2);
+
+    for (const tab of tabs) {
+      fireEvent.click(tab);
+      await waitFor(() => {
+        expect(getAllByTestId('project-card')).toHaveLength(1);
+        expect(getAllByTestId('issue-table')).toHaveLength(2);
+      });
+    }
+  });
+
+  it('renders total of 2 tabs (for 2 projects)', async () => {
+    const { getAllByRole } = await renderComponent();
+    const tabs = getAllByRole('tab');
+    expect(tabs).toHaveLength(2);
   });
 });