+++ plugins/readme/src/components/FetchComponent/FetchComponent.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { useEntity } from '@backstage/plugin-catalog-react';
 import { useApi } from '@backstage/core-plugin-api';
 import {