+++ plugins/readme/src/components/ReadmeCard/ReadmeCard.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import IconButton from '@mui/material/IconButton';
 import Box from '@mui/material/Box';
 import FullscreenIcon from '@mui/icons-material/Fullscreen';