+++ plugins/readme-backend/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @axis-backstage/plugin-readme-backend
 
+## 0.14.0
+
+### Minor Changes
+
+- 86c01b7: Bumped to Backstage 1.40
+
+## 0.13.1
+
+### Patch Changes
+
+- 733b2db: Added config to set the cache ttl.
+
 ## 0.13.0
 
 ### Minor Changes